+++ packages/enraiged/helpers/dates.php
@@ -1,27 +1,26 @@
 <?php
 
 /**
- *  avoid_holiday()
- *  Return a date that does not fall on a holiday or weekend.
- *  
- *  @param  mixed $stamp
- *  @param  mixed $format
+ *  Return a date that does not fall on a holiday or weekend in a specified country.
+ *
+ *  @param  string  $country
+ *  @param  mixed   $datetime
+ *  @param  string  $defer
  *  @return string
  */
-function avoid_holiday($stamp, $format = null)
+function avoid_holiday($country, $datetime, $defer = '+1 weekday')
 {
-    if (!preg_match("/^([0-9]+)$/", $stamp))
-    { $stamp = strtotime($stamp); }
+    if (!preg_match('/^([0-9]+)$/', $datetime))
+    { $datetime = strtotime($datetime); }
 
-    $stamp = (in_array(date("Y-m-d", $stamp), holidays()) || in_array(date("w", $stamp), [0,6]))
-        ? strtotime("+1 weekday", $stamp)
-        : $stamp;
+    $datetime = (in_array(date('Y-m-d', $datetime), holidays($country)) || in_array(date('w', $datetime), [0,6]))
+        ? strtotime($defer, $datetime)
+        : $datetime;
 
-    return ($format ? date($format, $stamp) : $stamp);
+    return $datetime;
 }
 
 /**
- *  datei18n()
  *  Return a simple preg_replace formatted date string.
  *
  *  @param  string  $datetime
@@ -42,17 +41,22 @@ function datei18n($datetime)
 }
 
 /**
- *  datetime()
  *  Return a formatted date/time, optionally from a provided datetime value.
- *  
- *  @param  mixed   $stamp = null
+ *
+ *  @param  mixed   $datetime = null
  *  @param  string  $format = 'Y-m-d H:i:s'
  *  @param  bool    $translate = false
  *  @return string
  */
-function datetime($stamp = null, $format = 'Y-m-d H:i:s', $translate = false)
+function datetime($datetime = null, $format = 'Y-m-d H:i:s', $translate = false)
 {
-    $formatted = date($format, timestamp($stamp ?? time()));
+    if (is_array($datetime)) {
+        return collect($datetime)
+            ->transform(fn ($each) => datetime($each, $format, $translate))
+            ->toArray();
+    }
+
+    $formatted = date($format, timestamp($datetime ?? time()));
 
     return $translate
         ? datei18n($formatted)
@@ -60,32 +64,30 @@ function datetime($stamp = null, $format = 'Y-m-d H:i:s', $translate = false)
 }
 
 /**
- *  datetimezone()
  *  Return a timezoned date/time, optionally from a provided datetime value.
  *
- *  @param  mixed   $stamp
+ *  @param  mixed   $datetime
  *  @param  string  $format
  *  @return string
  */
-function datetimezone($stamp = null, $format = 'Y-m-d H:i:s')
+function datetimezone($datetime = null, $format = 'Y-m-d H:i:s')
 {
-    return date($format, timezonestamp($stamp));
+    return date($format, timezonestamp($datetime));
 }
 
 /**
- *  elapsed()
  *  Return a human readable string representing an elapsed period of time.
- *  
- *  @param	mixed	$since
- *  @param	mixed	$stamp = 0
- *  @param	bool	$strict = true
- *  @return	string
+ *
+ *  @param  mixed   $since
+ *  @param  mixed   $datetime = 0
+ *  @param  bool    $strict = true
+ *  @return string
  */
-function elapsed($since, $stamp = 0, $strict = true)
+function elapsed($since, $datetime = 0, $strict = true)
 {
     $since = timestamp($since);
-    $stamp = $stamp ? timestamp($stamp) : time();
-    $lapse = ($stamp - $since);
+    $datetime = $datetime ? timestamp($datetime) : time();
+    $lapse = ($datetime - $since);
 
     $array = [
         ['diff' => 31536000, 'text' => __('year')],
@@ -100,7 +102,7 @@ function elapsed($since, $stamp = 0, $strict = true)
     $return = 'a moment';
 
     for ($i=0; $i<count($array); $i++) {
-        $item	= $array[$i];
+        $item = $array[$i];
 
         if ($lapse > $item['diff']) {
             $diff = floor($lapse / $item['diff']);
@@ -126,28 +128,70 @@ function elapsed($since, $stamp = 0, $strict = true)
 }
 
 /**
- *  holidays()
+ *  Return the array value of the furthest value in time, optionally from a specified timestamp.
+ *
+ *  @param  array   $array
+ *  @param  mixed   $from = null
+ *  @return mixed
+ */
+function furthest_datetime($array, $from = null)
+{
+    $index = furthest_datetime_index($array, $from = null);
+
+    return $array[$index];
+}
+
+/**
+ *  Return the array index of the furthest value in time, optionally from a specified timestamp.
+ *
+ *  @param  array   $array
+ *  @param  mixed   $from = null
+ *  @return int
+ */
+function furthest_datetime_index($array, $from = null)
+{
+    $from = $from ? timestamp($from) : time();
+    $furthest = null;
+    $index = null;
+
+    for ($i = 0; $i < count($array); $i++) {
+        $timestamp = timestamp($array[$i]);
+        $difference = $timestamp - $from;
+
+        if ($difference < 0) {
+            $difference = $difference *-1;
+        }
+
+        if (!$furthest || $difference > $furthest) {
+            $index = $i;
+        }
+    }
+
+    return $index;
+}
+
+/**
  *  Return an array of holiday dates for a country for a number of years.
- *  + $country is required as an iso alpha-2 country code; default: CA
- *  + $region is optional, not implemented at this time
- *  + $years will accept an array of years or a single year; default: current year plus one year
- *  
+ *
+ *  $country is required as an iso alpha-2 country code; default: CA
+ *  $region is optional, not implemented at this time
+ *  $years will accept an array of years or a single year; default: current year plus one year
+ *
  *  @param  string  $country = 'CA'
- *  @param  string  $region = null
  *  @param  array   $years = []
  *
  *  @todo   Regions? More countries?
  */
-function holidays($country = 'CA', $region = null, $years = [])
+function holidays($country = 'CA', $years = [])
 {
-    $dates	= [];
-    $years	= !is_array($years) ? [$years] : (!count($years) ? [date("Y"),date("Y")+1] : $years);
+    $dates = [];
+    $years = !is_array($years) ? [$years] : (!count($years) ? [date("Y") +0, date("Y") +1] : $years);
 
     foreach ($years as $year) {
         switch($country) {
-            // tested against: //https://www.officeholidays.com/countries/canada/{2018,2019}.php
+            // tested against: //https://www.officeholidays.com/countries/canada/2023
             case 'CA': $dates = array_merge($dates, [
-                date("Y-01-01"),
+                in_array(date("w", strtotime("jan 1 {$year}")), [0,6]) ? date("Y-m-d", strtotime("next monday", strtotime("jan 1 {$year}"))) : "{$year}-01-01",
                 date("Y-m-d", strtotime("{$year} february +3 monday")),
                 date("Y-m-d", strtotime("previous friday", easter_date($year))),
                 date("Y-m-d", strtotime("next monday", easter_date($year))),
@@ -158,7 +202,21 @@ function holidays($country = 'CA', $region = null, $years = [])
                 date("Y-m-d", strtotime("{$year} october +2 monday")),
                 in_array(date("w", strtotime("nov 11 {$year}")), [0,6]) ? date("Y-m-d", strtotime("next monday", strtotime("nov 11 {$year}"))) : "{$year}-11-11",
                 in_array(date("w", strtotime("dec 25 {$year}")), [0,6]) ? date("Y-m-d", strtotime("next monday", strtotime("dec 25 {$year}"))) : "{$year}-12-25",
-                "{$year}-12-26"
+                "{$year}-12-26",
+            ]); break;
+            // tested against: //https://www.officeholidays.com/countries/usa/2023
+            case 'US': $dates = array_merge($dates, [
+                in_array(date("w", strtotime("jan 1 {$year}")), [0,6]) ? date("Y-m-d", strtotime("next monday", strtotime("jan 1 {$year}"))) : "{$year}-01-01",
+                date("Y-01-01"),
+                date("Y-m-d", strtotime("{$year} january +3 monday")),
+                date("Y-m-d", strtotime("{$year} february +3 monday")),
+                (date("w", strtotime("may 31 {$year}")) == 1) ? "{$year}-05-31" : date("Y-m-d", strtotime("previous monday", strtotime("may 31 {$year}"))),
+                in_array(date("w", strtotime("jun 19 {$year}")), [0,6]) ? date("Y-m-d", strtotime("previous friday", strtotime("jun 19 {$year}"))) : "{$year}-06-19",
+                in_array(date("w", strtotime("jul 4 {$year}")), [0,6]) ? date("Y-m-d", strtotime("next monday", strtotime("jul 4 {$year}"))) : "{$year}-07-04",
+                date("Y-m-d", strtotime("{$year} september +1 monday")),
+                date("Y-m-d", strtotime("{$year} october +2 monday")),
+                date("Y-m-d", strtotime("{$year} november +4 thursday")),
+                in_array(date("w", strtotime("dec 25 {$year}")), [0,6]) ? date("Y-m-d", strtotime("next monday", strtotime("dec 25 {$year}"))) : "{$year}-12-25",
             ]); break;
         }
     }
@@ -167,18 +225,45 @@ function holidays($country = 'CA', $region = null, $years = [])
 }
 
 /**
- *  meridiem()
- *  Return a human readable string representing the general time of day.
- *  
- *  @param	string	$stamp
- *  @return	string
+ *  Determine if a provided date falls on a holiday.
+ *
+ *  @param  int|string  $date
+ *  @param  string  $country = 'CA'
+ *  @return bool
  */
-function meridiem($stamp)
+function is_holiday($date, $country = 'CA')
 {
-    $stamp = timestamp($stamp);
+    $timestamp = timestamp($date);
+    $year = date('Y', $timestamp);
 
-    $hour = date("Gi", $stamp);
-    $today = ($stamp > strtotime("today")) ? true : false;
+    return in_array(date('Y-m-d', $timestamp), holidays($country, $year));
+}
+
+/**
+ *  Determine if a provided date falls on a weekend.
+ *
+ *  @param  int|string  $date
+ *  @return bool
+ */
+function is_weekend($date)
+{
+    $timestamp = timestamp($date);
+
+    return in_array(date("w", $timestamp), [0,6]);
+}
+
+/**
+ *  Return a human readable string representing the general time of day.
+ *
+ *  @param  mixed   $datetime
+ *  @return string
+ */
+function meridiem($datetime)
+{
+    $timestamp = timestamp($datetime);
+
+    $hour = date("Gi", $timestamp);
+    $today = ($timestamp > strtotime("today")) ? true : false;
 
     $determiner = ($today ? 'this' : 'in the');
 
@@ -203,43 +288,102 @@ function meridiem($stamp)
 }
 
 /**
- *  next_business_day()
- *  Return the next consecutive business day after the provided date.
- *  
- *  @param  string  $stamp  Will accept a unix timestamp or php strtotime() argument
- *  @param  string  $format = null  Will accept a php strtotime() argument
- *  @return string
+ *  Return the array value of the nearest value in time, optionally from a specified timestamp.
+ *
+ *  @param  array   $array
+ *  @param  mixed   $from = null
+ *  @return mixed
  */
-function next_business_day($stamp, $format = null)
+function nearest_datetime($array, $from = null)
 {
-    $stamp = avoid_holiday( strtotime("+1 weekday", timestamp($stamp)) );
+    $index = nearest_datetime_index($array, $from = null);
 
-    return ($format ? date($format, $stamp) : $stamp);
+    return $array[$index];
 }
 
 /**
- *  next_interval()
+ *  Return the array index of the nearest value in time, optionally from a specified timestamp.
+ *
+ *  @param  array   $array
+ *  @param  mixed   $from = null
+ *  @return int
+ */
+function nearest_datetime_index($array, $from = null)
+{
+    $from = $from ? timestamp($from) : time();
+    $nearest = null;
+    $index = null;
+
+    for ($i = 0; $i < count($array); $i++) {
+        $timestamp = timestamp($array[$i]);
+        $difference = $timestamp - $from;
+
+        if ($difference < 0) {
+            $difference = $difference *-1;
+        }
+
+        if (!$nearest || $difference < $nearest) {
+            $nearest = $difference;
+            $index = $i;
+        }
+    }
+
+    return $index;
+}
+
+/**
+ *  Return the next consecutive business day from the provided date.
+ *
+ *  @param  string  $country
+ *  @param  mixed   $datetime  Will accept a unix timestamp or php strtotime() argument
+ *  @param  string  $format = null  Will accept a php strtotime() argument
+ *  @return string
+ */
+function next_business_day($country, $datetime, $format = null)
+{
+    $next_day = strtotime('+1 weekday', timestamp($datetime));
+    $timestamp = avoid_holiday($country, $next_day);
+
+    return ($format ? date($format, $timestamp) : $timestamp);
+}
+
+/**
  *  Return the timestamp of the next interval in minutes.
  *
  *  @param  int     $interval
- *  @param  mixed   $stamp = null
+ *  @param  mixed   $datetime = null
  *  @return int
  */
-function next_interval($interval, $stamp = null)
+function next_interval($interval, $datetime = null)
 {
-    $stamp = timestamp($stamp);
+    $timestamp = timestamp($timestamp);
 
-    $minutes = floor( date('i', $stamp) / $interval ) * $interval;
-    $flatten = timestamp(date("Y-m-d H:{$minutes}:00", $stamp));
+    $minutes = floor( date('i', $timestamp) / $interval ) * $interval;
+    $flatten = timestamp(date("Y-m-d H:{$minutes}:00", $timestamp));
     $inflate = strtotime("+{$interval} minutes", $flatten);
 
     return $inflate;
 }
 
 /**
- *  quarter_first_date()
+ *  Return the previous consecutive business day from the provided date.
+ *
+ *  @param  string  $country
+ *  @param  mixed   $datetime  Will accept a unix timestamp or php strtotime() argument
+ *  @param  string  $format = null  Will accept a php strtotime() argument
+ *  @return string
+ */
+function previous_business_day($country, $datetime, $format = null)
+{
+    $previous_day = strtotime('-1 weekday', timestamp($datetime));
+    $timestamp = avoid_holiday($country, $previous_day, '-1 weekday');
+
+    return ($format ? date($format, $timestamp) : $timestamp);
+}
+
+/**
  *  Return the first day of the current quarterly period.
- *  
+ *
  *  @param  string  $format = null
  *  @return string
  *
@@ -249,13 +393,12 @@ function quarter_first_date($format = null)
 {
     $Y = date('Y');
     $m = str_pad((ceil(date('m') /3) *3) -2, 2, "0", STR_PAD_LEFT);
-    $stamp = strtotime("{$Y}-{$m}-01");
+    $timestamp = strtotime("{$Y}-{$m}-01");
 
-    return ($format ? date($format, $stamp) : $stamp);
+    return ($format ? date($format, $timestamp) : $timestamp);
 }
 
 /**
- *  quarter_final_date()
  *  Return the final day of the current quarterly period
  *  
  *  @param  string  $format = null
@@ -266,27 +409,25 @@ function quarter_final_date($format = null)
     $Y = date('Y');
     $m = str_pad((ceil(date('m') /3) *3), 2, "0", STR_PAD_LEFT);
     $d = date('t', strtotime("{$Y}-{$m}"));
-    $stamp = strtotime("{$Y}-{$m}-{$d}");
+    $timestamp = strtotime("{$Y}-{$m}-{$d}");
 
-    return ($format ? date($format, $stamp) : $stamp);
+    return ($format ? date($format, $timestamp) : $timestamp);
 }
 
 /**
- *  timestamp()
  *  Return a unix timestamp equivalent of a provided php date.
- *  
- *  @param  string  $stamp
+ *
+ *  @param  mixed   $datetime = null
  *  @return int
  */
-function timestamp($stamp = null)
+function timestamp($datetime = null)
 {
-    return preg_match("/^([0-9]+)$/", $stamp ?: time())
-        ? $stamp ?: time()
-        : strtotime($stamp);
+    return preg_match("/^([0-9]+)$/", $datetime ?: time())
+        ? $datetime ?: time()
+        : strtotime($datetime);
 }
 
 /**
- *  timezone()
  *  Return the most user applicable timezone if possible, or default to alternate.
  *
  *  @return string
@@ -299,22 +440,20 @@ function timezone()
 }
 
 /**
- *  timezonestamp()
  *  Return a timezone converted unix timestamp from the provided datetime/timestamp.
  *
- *  @param  mixed   $stamp
- *  @param  string  $timezone
+ *  @param  mixed   $datetime = null
+ *  @param  string  $timezone = null
  *  @return mixed
  */
-function timezonestamp($stamp = null, $timezone = null)
+function timezonestamp($datetime = null, $timezone = null)
 {
-    $offset = timezone_offset_inverse();
+    $offset = timezone_offset_inverse($timezone);
 
-    return strtotime("{$offset} hrs", timestamp($stamp));
+    return strtotime("{$offset} hrs", timestamp($datetime));
 }
 
 /**
- *  timezone_offset()
  *  Return the +/- UTC offset integer of the provided (or default) timezone.
  *
  *  @param  string  $timezone = null
@@ -329,7 +468,6 @@ function timezone_offset($timezone = null)
 }
 
 /**
- *  timezone_offset_inverse()
  *  Return the +/- UTC offset inverse of the provided (or default) timezone.
  *
  *  @param  string  $timezone = null
@@ -341,4 +479,36 @@ function timezone_offset_inverse($timezone = null)
     $datetime = new \DateTime("now", $timezone);
 
     return sprintf("%+d", ($datetime->getOffset() /60 /60) *-1);
+}
+
+/**
+ *  Return a utc formatted date/time, optionally from a provided datetime value.
+ *
+ *  @param  mixed   $datetime = null
+ *  @param  string  $format = 'Y-m-d H:i:s'
+ *  @return string
+ */
+function utcdatetime($datetime = null, $format = 'Y-m-d H:i:s')
+{
+    if (is_array($datetime)) {
+        return collect($datetime)
+            ->transform(fn ($each) => utcdatetime($each, $format))
+            ->toArray();
+    }
+
+    return date($format, utcstamp($datetime));
+}
+
+/**
+ *  Return a utc timezone converted unix timestamp from the provided datetime/timestamp.
+ *
+ *  @param  mixed   $datetime = null
+ *  @param  string  $timezone = null
+ *  @return mixed
+ */
+function utcstamp($datetime = null, $timezone = null)
+{
+    $offset = timezone_offset($timezone);
+
+    return strtotime("{$offset} hrs", timestamp($datetime));
 }