+++ packages/enraiged/resources/js/components/forms/fields/DropdownField.vue
@@ -24,7 +24,7 @@
                 :placeholder="placeholder"
                 :show-clear="field.clearable || clearable"
                 @filter="filter"
-                @update:modelValue="update"/>
+                @update:modelValue="update; $emit('update:modelValue', $event)"/>
             <div class="error p-error" v-if="error">
                 <i class="pi pi-exclamation-circle" v-tooltip.top="error"></i>
                 <span class="message">{{ error }}</span>