+++ packages/enraiged/resources/js/components/forms/fields/TextareaField.vue
@@ -18,7 +18,7 @@
                 :id="id"
                 :placeholder="placeholder"
                 :rows="field.rows || rows"
-                @update:modelValue="update"/>
+                @update:modelValue="update; $emit('update:modelValue', $event)"/>
             <div class="error p-error" v-if="error">
                 <i class="pi pi-exclamation-circle" v-tooltip.top="error"></i>
                 <span class="message">{{ error }}</span>