+++ packages/enraiged/src/Support/Database/Traits/CreatedBy.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Enraiged\Support\Database\Traits;
+
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Support\Facades\Auth;
+
+trait CreatedBy
+{
+    /**
+     *  @return void
+     */
+    public static function bootCreatedBy()
+    {
+        self::creating(fn ($model) => $model->setCreatedBy());
+    }
+
+    /**
+     *  @return \Illuminate\Database\Eloquent\Relations\BelongsTo
+     */
+    public function createdBy(): BelongsTo
+    {
+        return $this->belongsTo(config('auth.providers.users.model'), 'created_by');
+    }
+
+    /**
+     *  @return void
+     */
+    private function setCreatedBy()
+    {
+        $this->created_by = Auth::check() ? Auth::id() : 1;
+    }
+}